% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load.model}
\alias{load.model}
\title{Load the classification model.}
\usage{
load.model(mtype = "gbm")
}
\arguments{
\item{mtype}{Type of the model to be returned. Valid value [svm | rf | gbm]. Defaults to the highest accuracy gbm model.}
}
\value{
a caret::train binary classification model object trained using specified model types.
}
\description{
Loads a pre-trained persistent binary classification model to classify customers as "PHOTOGRAPHERS" or "OTHERS".
}
\examples{
    load.model()
    load.model(mtype="rf")
    
}
